+++ src/main/java/com/laoqi/assistant/entity/MessageEntity.java
@@ -8,17 +8,17 @@ import com.baomidou.mybatisplus.annotation.TableName;
 public class MessageEntity {
 
     @TableId(type = IdType.AUTO)
-    private Integer id;
+    private Long id;
     private String sessionId;
     private String source;
     private String role;
     private String content;
     private String mode;
-    private Integer kbId;
+    private Long kbId;
     private String createdAt;
 
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
     public String getSessionId() { return sessionId; }
     public void setSessionId(String sessionId) { this.sessionId = sessionId; }
     public String getSource() { return source; }
@@ -29,8 +29,8 @@ public class MessageEntity {
     public void setContent(String content) { this.content = content; }
     public String getMode() { return mode; }
     public void setMode(String mode) { this.mode = mode; }
-    public Integer getKbId() { return kbId; }
-    public void setKbId(Integer kbId) { this.kbId = kbId; }
+    public Long getKbId() { return kbId; }
+    public void setKbId(Long kbId) { this.kbId = kbId; }
     public String getCreatedAt() { return createdAt; }
     public void setCreatedAt(String createdAt) { this.createdAt = createdAt; }
 }
