+++ src/main/java/com/laoqi/assistant/entity/SessionEntity.java
@@ -15,12 +15,12 @@ public class SessionEntity {
     private String mode;
     private String createdAt;
     private String updatedAt;
-    private Integer kbId;
+    private Long kbId;
 
     public String getId() { return id; }
     public void setId(String id) { this.id = id; }
-    public Integer getKbId() { return kbId; }
-    public void setKbId(Integer kbId) { this.kbId = kbId; }
+    public Long getKbId() { return kbId; }
+    public void setKbId(Long kbId) { this.kbId = kbId; }
 
     public String getSource() { return source; }
     public void setSource(String source) { this.source = source; }
