+++ src/main/java/com/laoqi/assistant/service/db/MessageDbService.java
@@ -9,7 +9,7 @@ public interface MessageDbService extends IService<MessageEntity> {
     List<MessageEntity> listBySession(String sessionId);
     List<MessageEntity> listBySessionAndMode(String sessionId, String mode);
     List<MessageEntity> listRecentBySession(String sessionId, int limit);
-    List<MessageEntity> listByKb(Integer kbId, int offset, int limit);
-    long countByKb(Integer kbId);
-    List<MessageEntity> searchByKb(Integer kbId, String q, int limit);
+    List<MessageEntity> listByKb(Long kbId, int offset, int limit);
+    long countByKb(Long kbId);
+    List<MessageEntity> searchByKb(Long kbId, String q, int limit);
 }
